datapath: move PreFilter to pkg/datapath/prefilter.PreFilter
